# core-lending-asset-finance-sdk/src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:18080
  description: Generated server url
...
paths:
...
  /api/v1/asset-finance-agreements/{agreementId}/assets/{assetId}/return-records/{recordId}:
    get:
      tags:
      - ReturnRecord
      summary: Get a return record by ID
      description: Retrieve a specific return record by its unique identifier
      operationId: getById_5
      parameters:
      - name: agreementId
        in: path
        description: Unique identifier of the agreement
        required: true
        schema:
          type: string
          format: uuid
      - name: assetId
        in: path
        description: Unique identifier of the asset
        required: true
        schema:
          type: string
          format: uuid
      - name: recordId
        in: path
        description: Unique identifier of the return record
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Return record found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnRecordDTO"
        "404":
          description: "Return record, asset, or agreement not found"
        "500":
          description: Internal server error
    put:
      tags:
      - ReturnRecord
      summary: Update an existing return record
      description: Update all fields of an existing return record
      operationId: update_5
      parameters:
      - name: agreementId
        in: path
        description: Unique identifier of the agreement
        required: true
        schema:
          type: string
          format: uuid
      - name: assetId
        in: path
        description: Unique identifier of the asset
        required: true
        schema:
          type: string
          format: uuid
      - name: recordId
        in: path
        description: Unique identifier of the return record to update
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReturnRecordDTO"
        required: true
      responses:
        "200":
          description: Return record updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnRecordDTO"
        "400":
          description: Invalid input data
        "404":
          description: "Return record, asset, or agreement not found"
        "500":
          description: Internal server error
    delete:
      tags:
      - ReturnRecord
      summary: Delete a return record
      description: Permanently delete a return record
      operationId: delete_5
      parameters:
      - name: agreementId
        in: path
        description: Unique identifier of the agreement
        required: true
        schema:
          type: string
          format: uuid
      - name: assetId
        in: path
        description: Unique identifier of the asset
        required: true
        schema:
          type: string
          format: uuid
      - name: recordId
        in: path
        description: Unique identifier of the return record to delete
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "204":
          description: Return record deleted successfully
        "404":
          description: "Return record, asset, or agreement not found"
        "500":
          description: Internal server error
...
  /api/v1/asset-finance-agreements/{agreementId}/assets/{assetId}/return-records:
    get:
      tags:
      - ReturnRecord
      summary: List/Search return records for an asset
      description: Retrieve a paginated list of return/condition assessment records
        for a specific asset
      operationId: findAll_5
      parameters:
      - name: agreementId
        in: path
        description: Unique identifier of the agreement
        required: true
        schema:
          type: string
          format: uuid
      - name: assetId
        in: path
        description: Unique identifier of the asset
        required: true
        schema:
          type: string
          format: uuid
      - name: filterRequest
        in: query
        description: Filter criteria for searching return records
        required: true
        schema:
          $ref: "#/components/schemas/FilterRequestReturnRecordDTO"
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Successfully retrieved list of return records
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaginationResponseReturnRecordDTO"
        "400":
          description: Invalid filter request
        "404":
          description: Asset or agreement not found
        "500":
          description: Internal server error
    post:
      tags:
      - ReturnRecord
      summary: Create a new return record for an asset
      description: Create a new return record to document asset condition assessment
        upon return
      operationId: create_5
      parameters:
      - name: agreementId
        in: path
        description: Unique identifier of the agreement
        required: true
        schema:
          type: string
          format: uuid
      - name: assetId
        in: path
        description: Unique identifier of the asset
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReturnRecordDTO"
        required: true
      responses:
        "200":
          description: Return record created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnRecordDTO"
        "400":
          description: Invalid input data
        "404":
          description: Asset or agreement not found
        "500":
          description: Internal server error
...
components:
  schemas:
...
    ReturnRecordDTO:
      required:
      - actualReturnDate
      - assetFinanceAssetId
      - conditionReport
      type: object
      properties:
        returnRecordId:
          type: string
          format: uuid
          readOnly: true
        assetFinanceAssetId:
          type: string
          format: uuid
        actualReturnDate:
          type: string
          format: date
        conditionReport:
          maxLength: 2000
          minLength: 0
          type: string
        damageCost:
          type: number
        isFinalized:
          type: boolean
        note:
          maxLength: 1000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: "The direction of sorting, either ASC or DESC."
          example: DESC
          default: DESC
      description: "Represents a pagination request for retrieving paginated results,\
        \ including page number, size, sort field, and direction."
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/RangeObject"
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
...
    FilterRequestReturnRecordDTO:
      required:
      - pagination
      type: object
      properties:
        filters:
          $ref: "#/components/schemas/ReturnRecordDTO"
        rangeFilters:
          $ref: "#/components/schemas/RangeFilter"
        pagination:
          $ref: "#/components/schemas/PaginationRequest"
        options:
          $ref: "#/components/schemas/FilterOptions"
      description: Generic filter request that includes both filter criteria and pagination
    PaginationResponseReturnRecordDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: "#/components/schemas/ReturnRecordDTO"
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: "The current page number, typically zero-based."
          format: int32
      description: Represents a paginated response containing a list of items and
        pagination metadata.
...